requirements.txt
fastapi
uvicorn
pandas
numpy
scipy
statsmodels
scikit-learn
matplotlib
seaborn
python-dotenv
supabase
psycopg2-binary
ipykernel
jupyterlab
